Add test name to parameters
